requirements.txt
python-dotenv 
notebook 
langgraph
langchain-openai
tavily-python 
langchain_community
langchain-tavily
requests
streamlit
langgraph-supervisor
langchain-mcp-adapters 
fastmcp
amadeus